+++ python/lsst/ts/scheduler/utils/test/feature_scheduler_sim.py
@@ -28,9 +28,8 @@ from astropy.time import Time
 from lsst.ts.astrosky.model import AstronomicalSkyModel
 from lsst.ts.dateloc import ObservatoryLocation
 from lsst.ts.observatory.model import ObservatoryModel, ObservatoryState
-from rubin_sim.site_models.cloudModel import CloudModel
-from rubin_sim.site_models.downtimeModel import DowntimeModel
-from rubin_sim.site_models.seeingModel import SeeingModel
+from rubin_sim.site_models.cloud_model import CloudModel
+from rubin_sim.site_models.seeing_model import SeeingModel
 
 from ...driver import FeatureScheduler
 from ...driver.feature_scheduler_target import FeatureSchedulerTarget
@@ -82,7 +81,6 @@ class FeatureSchedulerSim:
         self.models["sky"].sky_brightness_pre.load_length = 7
         self.models["seeing"] = SeeingModel()
         self.models["cloud"] = CloudModel()
-        self.models["downtime"] = DowntimeModel()
 
         self.driver = FeatureScheduler(
             models=self.models, raw_telemetry=self.raw_telemetry